+++ src/commands/discord/Invite.ts
@@ -1,4 +1,4 @@
-import { ChatInputCommandInteraction, OAuth2Scopes, PermissionsBitField } from 'discord.js';
+import { ChatInputCommandInteraction, PermissionsBitField } from 'discord.js';
 import Command from '../../base/classes/Command';
 import CustomClient from '../../base/classes/CustomClient';
 import Category from '../../base/enums/Category';
@@ -18,19 +18,7 @@ export default class Invite extends Command {
   }
 
   async Execute(interaction: ChatInputCommandInteraction) {
-    const inviteLink = this.client.generateInvite({
-      scopes: [OAuth2Scopes.Bot, OAuth2Scopes.ApplicationsCommands],
-      permissions: [
-        PermissionsBitField.Flags.SendMessages,
-        PermissionsBitField.Flags.EmbedLinks,
-        PermissionsBitField.Flags.AttachFiles,
-        PermissionsBitField.Flags.ReadMessageHistory,
-        PermissionsBitField.Flags.UseApplicationCommands,
-        PermissionsBitField.Flags.ManageThreads,
-        PermissionsBitField.Flags.CreatePrivateThreads,
-        PermissionsBitField.Flags.SendMessagesInThreads,
-      ],
-    });
+    const inviteLink = this.client.GetInviteLink();
 
     await interaction.reply({
       content: `Here is the invite link for the bot: [Invite Link](${inviteLink})`,